actualizacion

+++ 08-laravel/gameapp/database/migrations/2024_06_14_014351_alter_users_table.php
@@ -4,29 +4,51 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
-     * Run the migrations.
+     * Ejecuta las migraciones.
      */
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->renameColumn('name', 'fullname');
-            $table->string('gender')->after('name');
-            $table->string('role')->default('Customer')->after('password');
+            // Verifica si la columna 'name' existe antes de renombrarla
+            if (Schema::hasColumn('users', 'name')) {
+                $table->renameColumn('name', 'fullname');
+            }
+
+            // Verifica si la columna 'gender' no existe antes de agregarla
+            if (!Schema::hasColumn('users', 'gender')) {
+                $table->string('gender')->nullable()->after('fullname');
+            }
+
+            // Verifica si la columna 'role' no existe antes de agregarla
+            if (!Schema::hasColumn('users', 'role')) {
+                $table->string('role')->default('Customer')->after('password');
+            }
         });
     }
 
     /**
-     * Reverse the migrations.
+     * Reversa las migraciones.
      */
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->renameColumn('fullname', 'name');
-            $table->dropColumn('gender');
-            $table->dropColumn('role');
+            // Revertir el nombre de la columna 'fullname' a 'name'
+            if (Schema::hasColumn('users', 'fullname')) {
+                $table->renameColumn('fullname', 'name');
+            }
+
+            // Elimina las columnas 'gender' y 'role' si existen
+            if (Schema::hasColumn('users', 'gender')) {
+                $table->dropColumn('gender');
+            }
+            if (Schema::hasColumn('users', 'role')) {
+                $table->dropColumn('role');
+            }
         });
     }
 };
+
